+++ Cargo.toml
@@ -17,7 +17,6 @@ version = "0.151.1"
 
 [workspace.dependencies]
 # Local dependencies
-<<<<<<< HEAD
 boojum-cuda = { version = "=0.151.0", path = "crates/boojum-cuda" }
 era_criterion_cuda = { version = "=0.151.0", path = "crates/criterion-cuda" }
 era_cudart = { version = "=0.151.0", path = "crates/cudart" }
@@ -30,18 +29,6 @@ shivini = { version = "=0.151.0", path = "crates/shivini" }
 wrapper-prover = { version = "=0.151.0", path = "crates/wrapper-prover", package = "zksync-wrapper-prover" }
 fflonk = { version = "=0.151.0", path = "crates/fflonk", package = "fflonk-gpu" }
 proof-compression = { version = "=0.150.9", path = "crates/proof-compression", package = "proof-compression" }
-=======
-boojum-cuda = { version = "=0.151.1", path = "crates/boojum-cuda" }
-era_criterion_cuda = { version = "=0.151.1", path = "crates/criterion-cuda" }
-era_cudart = { version = "=0.151.1", path = "crates/cudart" }
-era_cudart_sys = { version = "=0.151.1", path = "crates/cudart-sys" }
-era_cudart_sys_bindings_generator = { version = "=0.151.1", path = "crates/cudart-sys-bindings-generator" }
-gpu-ffi = { version = "=0.151.1", path = "crates/gpu-ffi", package = "zksync-gpu-ffi" }
-gpu-ffi-bindings-generator = { version = "=0.151.1", path = "crates/gpu-ffi", package = "zksync-gpu-ffi-bindings-generator" }
-gpu-prover = { version = "=0.151.1", path = "crates/gpu-prover", package = "zksync-gpu-prover" }
-shivini = { version = "=0.151.1", path = "crates/shivini" }
-wrapper-prover = { version = "=0.151.1", path = "crates/wrapper-prover", package = "zksync-wrapper-prover" }
->>>>>>> e22c918 (chore(main): release 0.151.1 (#40))
 
 # These dependencies should be shared by all the crates.
 # circuit_definitions = { version = "=0.150.9" }
